Add the straight flush hand requirement.
Notes about decisions:

1. Used london style / mocking tests because the straight flush is exactly
a straight and exactly a flush -- to test it chicago style I would have
had to replicate all of the tests for both. Since I wanted to do a full
truth table:

        !straight + !flush => false
        !straight + flush  => false
        straight + !flush  => false
        straight + flush   => match

That would have been a lot to test.

2. Pass a clone of the hand instead of the hand itself:
I know from the implementation of those requirements that they don't
mutate the hand they are passed, but there's no mechanism to enforce
that. Better safe than sorry.