Move fixtures for integration test to top level directory
Also slightly improve the test message for those integration tests.

The examples for the integration tests serve as a good documentation for
what scenarios I have covered. It would probably be helfpul for
a reviewer to see that in the top level directory. Skimming through it
should give them a quick overview of what this implementation does (and
doesn't?) do.