Add a separate tsconfig for build/test so they both work
When I added the cli (and updated the tsconfig) the tests stopped
working.

If this were a longer-running project I would take more care to move as
much as possible to the base tsconfig -- I am sure that there are some
things in the tsconfig.cli that wouldn't break the tests if they were
moved to the base tsconfig. Alas, it's not worth the trouble at this
point.