Remove company name from the code
Yes, it's still here in the commit history -- but at least the landing page won't have it.

I wonder if Google indexes the previous commits? I'd guess not, but who knows?

I'm not up for completely scrubbing the commit history to remove it, even though that's
probably the best thing to do?